# threadset: revise/simplify on top of `GAsyncQueue`