--- a/ChangeLog
+++ b/ChangeLog
@@ -7,6 +7,7 @@
 - threadpool: fix a race condition in error handling [kleisauke]
 - disable GLib cast checks and asserts for plain builds [kleisauke]
 - fix jpeg in tiff for high Q [nahilsobh]
+- threadset: fix a race condition during thread exit [kleisauke]
 
 11/8/24 8.15.3