--- a/libvips/include/vips/internal.h
+++ b/libvips/include/vips/internal.h
@@ -151,7 +151,6 @@ extern gboolean vips__cache_trace;
 void vips__thread_init(void);
 void vips__threadpool_init(void);
 void vips__threadpool_shutdown(void);
-int vips__thread_execute(const char *name, GFunc func, gpointer data);
 VIPS_API void vips__worker_lock(GMutex *mutex);
 
 void vips__cache_init(void);