The previous implementation was prone to race conditions under rare circumstances.

Targets the 8.15 branch.

Context: #4069.