@kleisauke 2024-09-06
Marked as draft due to this note: https://github.com/libvips/libvips/discussions/4069#discussioncomment-10569605.