@kleisauke 2024-09-06
FWIW, CIFuzz build will be fixed after PR https://github.com/google/oss-fuzz/pull/12463 lands.