@kleisauke 2024-09-21
The [`weserv/images`](https://github.com/weserv/images) project has been running this implementation without any issues for some time. Additionally, I didn't observe any perf regressions in the https://github.com/kleisauke/vips-microbench repo with this implementation.

<details>
  <summary>Details</summary>

Before:
```console
$ ./bin/vips-microbench-jpeg --benchmark_repetitions=10
2024-09-21T12:28:55+02:00
Running ./bin/vips-microbench-jpeg
Run on (24 X 5105.97 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x12)
  L1 Instruction 32 KiB (x12)
  L2 Unified 1024 KiB (x12)
  L3 Unified 32768 KiB (x2)
Load Average: 0.84, 1.26, 1.31
---------------------------------------------------------------------------------------------------
Benchmark                                                         Time             CPU   Iterations
---------------------------------------------------------------------------------------------------
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time               445 ms         54.0 ms            2
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time               444 ms         51.4 ms            2
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time               438 ms         51.2 ms            2
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time               435 ms         51.4 ms            2
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time               428 ms         49.4 ms            2
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time               435 ms         52.6 ms            2
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time               430 ms         49.8 ms            2
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time               428 ms         50.4 ms            2
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time               429 ms         51.4 ms            2
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time               429 ms         51.7 ms            2
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time_mean          434 ms         51.3 ms           10
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time_median        432 ms         51.4 ms           10
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time_stddev       6.57 ms         1.33 ms           10
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time_cv           1.51 %          2.58 %            10
```

After:
```console
$ ./bin/vips-microbench-jpeg --benchmark_repetitions=10
2024-09-21T12:32:13+02:00
Running ./bin/vips-microbench-jpeg
Run on (24 X 400 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x12)
  L1 Instruction 32 KiB (x12)
  L2 Unified 1024 KiB (x12)
  L3 Unified 32768 KiB (x2)
Load Average: 0.78, 1.07, 1.23
---------------------------------------------------------------------------------------------------
Benchmark                                                         Time             CPU   Iterations
---------------------------------------------------------------------------------------------------
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time               451 ms         51.5 ms            2
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time               428 ms         52.0 ms            2
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time               434 ms         51.0 ms            2
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time               440 ms         52.2 ms            2
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time               434 ms         51.8 ms            2
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time               436 ms         53.3 ms            2
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time               436 ms         51.1 ms            2
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time               430 ms         50.3 ms            2
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time               432 ms         51.5 ms            2
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time               429 ms         51.9 ms            2
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time_mean          435 ms         51.7 ms           10
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time_median        434 ms         51.7 ms           10
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time_stddev       6.78 ms        0.792 ms           10
BM_jpeg_crop_shrink_sharpen/10000x10000/real_time_cv           1.56 %          1.53 %            10
```
</details>

So, this is ready for review now. If necessary, this PR could target the master branch instead.