@jcupitt 2024-10-02
I read it carefully and it seems fine. Thanks for doing all this work!

It'd be good to understand why the old implementation raced :( it feels bad to be replacing something without knowing why.