@jcupitt 2024-10-02
Ah OK, thanks for the explanation. Threaded programming is hard :(