@kleisauke 2024-10-02
> So far, it's performing well, using approximately the same number of libvips workers when queried with:
> 
> ```console
> $ ps -eLo comm | grep "libvips worker" | wc -l
> ```

Actually, the number of threads is much lower on my 6-core test server under roughly the same load. For reference:

Before:
```console
$ ps -eLo comm | grep "libvips worker" | wc -l
516
```

After:
```console
$ ps -eLo comm | grep "libvips worker" | wc -l
60
```

This is because we allow a maximum of 2 idle threads:
https://github.com/libvips/libvips/blob/9b126ddb9499ad94c585a11e04416a5de554b6fd/libvips/iofuncs/threadset.c#L96-L98
(previously this was not constrained)